refactor: JMH benchmarks, dont use builtin
